% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lpbenchmark.R
\name{Lpbenchmark}
\alias{Lpbenchmark}
\title{Benchmarking Suite}
\usage{
Lpbenchmark(size = 1e+08, test_threads = seq_len(parallel::detectCores()),
  repeats = 5, seed = 1, html_out = TRUE,
  output_file = "benchmark.html", output_dir = "./", open_file = FALSE)
}
\arguments{
\item{size}{The size of the benchmark. If you don't know what value to use. Defaults to \code{1e8} (approximately 750MB, requires at least 750MB x 4 = 3GB).}

\item{test_threads}{The number of threads to test, to provide as a vector. Defaults to \code{seq_len(parallel::detectCores())} which means try from 1 to all threads.}

\item{repeats}{The number of times to retry benchmarking to average out results. Defaults to \code{5}.}

\item{seed}{The random number generator seed. It should not have any impact on performance. Defaults to \code{1}.}

\item{html_out}{Should the benchmark be done using a markdown file for HTML output? Defaults to \code{TRUE}.}

\item{output_file}{The output file. Defaults to \code{"benchmark.html"}.}

\item{output_dir}{The output directory for the output file. Defaults to \code{"./"}.}

\item{open_file}{Open the file at the end of the benchmark. Defaults to \code{FALSE}.}
}
\value{
A markdown file.
}
\description{
Runs the integrated benchmarking suite.
}
\examples{
\dontrun{
# Output benchmark results with 1 to 2 threads in console
Lpbenchmark(size = 1e7, test_threads = 1:2, repeats = 1, html_out = FALSE)

# Output benchmark repeated 5 times using 1 to all threads in HTML file and opens it immediately
Lpbenchmark(size = 1e7, repeats = 5, html_out = TRUE,
            output_file = tempdir(), output_dir = tempdir(), open_file = TRUE)
}

}
